+++ R/manytrade-memberships.R
@@ -5,11 +5,17 @@
 #' For more information and references to each of the datasets used,
 #' please use the `data_source()` and `data_contrast()` functions.
 #'\describe{
-#' \item{GPTAD_MEM: }{A dataset with 2221 observations and the following
-#' 9 variables: many_ID, GPTAD_ID, Country_ID, Country, Title, Beg, Signature, 
-#' Force, treaty_ID.}
-#' \item{DESTA_MEM: }{A dataset with 7772 observations and the following
-#' 8 variables: many_ID, DESTA_ID, Country, Title, Beg, Signature, Force, 
-#' treaty_ID.}
+#' \item{GPTAD_MEM: }{A dataset with 2201 observations and the following
+#' 9 variables: manyID, CountryID, Title, Beg, Signature, Force,
+#' CountryName, gptadID, treatyID.}
+#' \item{DESTA_MEM: }{A dataset with 7492 observations and the following
+#' 9 variables: manyID, CountryID, Title, Beg, Signature, Force,
+#' CountryName, destaID, treatyID.}
 #' }
- "memberships"
+
+#'
+#' @details
+#' ``` {r, echo = FALSE, warning = FALSE}
+#' lapply(memberships, skimr::skim_without_charts)
+#' ```
+"memberships"