+++ R/manytrade-references.R
@@ -6,6 +6,12 @@
 #' please use the `data_source()` and `data_contrast()` functions.
 #'\describe{
 #' \item{DESTA_REF: }{A dataset with 1000 observations and the following
-#' 3 variables: many_ID1, RefType, many_ID2.}
+#' 3 variables: manyID1, RefType, manyID2.}
 #' }
- "references"
+
+#'
+#' @details
+#' ``` {r, echo = FALSE, warning = FALSE}
+#' lapply(references, skimr::skim_without_charts)
+#' ```
+"references"