+++ data-raw/memberships/GPTAD_MEM/prepare-GPTAD_MEM.R
@@ -11,72 +11,83 @@ GPTAD_MEM <- read.csv("data-raw/memberships/GPTAD_MEM/GPTAD.csv")
 # formats of the 'GPTAD_MEM' object until the object created
 # below (in stage three) passes all the tests.
 GPTAD_MEM <- as_tibble(GPTAD_MEM) %>%
-  dplyr::mutate(GPTAD_ID = dplyr::row_number()) %>%
-  dplyr::mutate(Country = gsub("\\\\r\\\\n", "", Membership)) %>%
+  dplyr::mutate(gptadID = as.character(dplyr::row_number())) %>%
+  dplyr::mutate(CountryName = gsub("\\\\r\\\\n", "", Membership)) %>%
   #remove \r\n line break entries
-  dplyr::mutate(Country = stringr::str_replace(Country, "China (Taiwan), Nicaragua", "Taiwan, Nicaragua")) %>%
+  dplyr::mutate(CountryName = stringr::str_replace(CountryName, "China (Taiwan), Nicaragua", "Taiwan, Nicaragua")) %>%
   #standardize delimitation of values by commas
-  dplyr::mutate(Country = stringr::str_replace(Country, "St. Kitts-Nevis-Anguilla", "Saint Kitts and Nevis, Anguilla")) %>%
-  dplyr::mutate(Country = stringr::str_replace(Country, "Singapore - Korea", "Singapore, Korea")) %>%
-  dplyr::mutate(Country = gsub("[()]", ",", Country)) %>%
-  dplyr::mutate(Country = gsub("[;]", ",", Country)) %>%
-  dplyr::mutate(Country = gsub("[.]", "", Country)) %>%
-  dplyr::mutate(Country = stringr::str_replace(Country, "Thailand Vietnam", "Thailand, Vietnam")) %>% #add comma between country names
-  dplyr::mutate(Country = stringr::str_replace(Country, "Finland France", "Finland, France")) %>%
-  dplyr::mutate(Country = stringr::str_replace(Country, "Paraguay and Uruguay", "Paraguay, Uruguay")) %>%
-  dplyr::mutate(Country = stringr::str_replace(Country, "Russian Federation and Ukraine", "Russian Federation, Ukraine")) %>% 
-  dplyr::mutate(Country = stringr::str_replace(Country, "Slovak Republic and Slovenia", "Slovak Republic, Slovenia")) %>%
-  dplyr::mutate(Country = stringr::str_replace(Country, "Spain Sweden", "Spain, Sweden")) %>%
-  dplyr::mutate(Country = stringr::str_replace(Country, "Ukraine and Uzbekistan", "Ukraine, Uzbekistan")) %>%
-  dplyr::mutate(Country = stringr::str_replace(Country, "Uzbekistan and Ukraine", "Ukraine, Uzbekistan")) %>%
-  dplyr::mutate(Country = stringr::str_replace(Country, "Suriname and Trinidad and Tobago", "Suriname, Trinidad and Tobago")) %>%
-  dplyr::mutate(Country = stringr::str_replace(Country, "Norway and Switzerland", "Norway, Switzerland")) %>%
+  dplyr::mutate(CountryName = stringr::str_replace(CountryName, "St. Kitts-Nevis-Anguilla", "Saint Kitts and Nevis, Anguilla")) %>%
+  dplyr::mutate(CountryName = stringr::str_replace(CountryName, "Singapore - Korea", "Singapore, Korea")) %>%
+  dplyr::mutate(CountryName = gsub("[()]", ",", CountryName)) %>%
+  dplyr::mutate(CountryName = gsub("[;]", ",", CountryName)) %>%
+  dplyr::mutate(CountryName = gsub("[.]", "", CountryName)) %>%
+  #add comma between country names
+  dplyr::mutate(CountryName = stringr::str_replace(CountryName, "Thailand Vietnam", "Thailand, Vietnam")) %>% 
+  dplyr::mutate(CountryName = stringr::str_replace(CountryName, "Finland France", "Finland, France")) %>%
+  dplyr::mutate(CountryName = stringr::str_replace(CountryName, "Paraguay and Uruguay", "Paraguay, Uruguay")) %>%
+  dplyr::mutate(CountryName = stringr::str_replace(CountryName, "Russian Federation and Ukraine", "Russian Federation, Ukraine")) %>% 
+  dplyr::mutate(CountryName = stringr::str_replace(CountryName, "Slovak Republic and Slovenia", "Slovak Republic, Slovenia")) %>%
+  dplyr::mutate(CountryName = stringr::str_replace(CountryName, "Spain Sweden", "Spain, Sweden")) %>%
+  dplyr::mutate(CountryName = stringr::str_replace(CountryName, "Ukraine and Uzbekistan", "Ukraine, Uzbekistan")) %>%
+  dplyr::mutate(CountryName = stringr::str_replace(CountryName, "Uzbekistan and Ukraine", "Ukraine, Uzbekistan")) %>%
   #remove 'and' between country names
-  tidyr::separate_rows(Country, sep=",") %>%
+  dplyr::mutate(CountryName = stringr::str_replace(CountryName, "Suriname and Trinidad and Tobago", "Suriname, Trinidad and Tobago")) %>%
+  dplyr::mutate(CountryName = stringr::str_replace(CountryName, "Norway and Switzerland", "Norway, Switzerland")) %>%
+  tidyr::separate_rows(CountryName, sep=",") %>%
   #separate column into rows by each country
-  dplyr::mutate(Country = gsub("^and | and$", "", Country)) %>%
-  dplyr::mutate(Country = ifelse(Country == "", NA, Country)) %>%
+  dplyr::mutate(CountryName = gsub("^and | and$", "", CountryName)) %>%
+  dplyr::mutate(CountryName = ifelse(CountryName == "", NA, CountryName)) %>%
   #standardize empty rows and 
-  dplyr::mutate(Country = ifelse(Country == " ", NA, Country)) %>%
+  dplyr::mutate(CountryName = ifelse(CountryName == " ", NA, CountryName)) %>%
   #rows with only whitespace to NA
-  dplyr::filter(Country != "NA") %>% #remove missing rows
-  dplyr::mutate(Country = trimws(Country, which = c("both", "left", "right"), 
+  dplyr::filter(CountryName != "NA") %>% #remove missing rows
+  dplyr::mutate(CountryName = trimws(CountryName, which = c("both", "left", "right"), 
                                  whitespace = "[ \t\r\n]")) %>%
   #remove whitespace
-  dplyr::mutate(Country = ifelse(Country == "", NA, Country)) %>%
-  dplyr::filter(Country != "NA",
-                Country != "British") %>%
+  dplyr::mutate(CountryName = ifelse(CountryName == "", NA, CountryName)) %>%
+  dplyr::filter(CountryName != "NA",
+                CountryName != "British") %>%
   #remove empty rows and redundant 'British' in data
-  dplyr::mutate(Country = manytrade::code_countryname(Country)) %>%
+  dplyr::mutate(CountryName = manystates::code_states(CountryName)) %>%
   #translate French country names and correct spelling
-  dplyr::mutate(Country = dplyr::recode(Country, "EC" = "European Community")) %>%
+  dplyr::mutate(CountryName = dplyr::recode(CountryName, "EC" = "European Community")) %>%
   #not included in regex list because of overlaps with other country names
-  dplyr::mutate(Country_ID = countrycode::countrycode(Country, 
-                                                      origin = 'country.name',
-                                                      destination = 'iso3n')) %>%
+  dplyr::mutate(CountryID = countrycode::countrycode(CountryName, 
+                                                     origin = 'country.name',
+                                                     destination = 'iso3c')) %>%
   #add iso code for country names
   dplyr::mutate(`Date.of.Signature` = ifelse(`Date.of.Signature`=="n/a", 
                                              NA, `Date.of.Signature`)) %>%
-  dplyr::mutate(`Date.of.Entry.into.Force` = ifelse(`Date.of.Entry.into.Force`=="N/A", NA, `Date.of.Entry.into.Force`)) %>%
+  dplyr::mutate(`Date.of.Entry.into.Force` = ifelse(`Date.of.Entry.into.Force`=="N/A", 
+                                                    NA, `Date.of.Entry.into.Force`)) %>%
   manydata::transmutate(Title = manypkgs::standardise_titles(`Common.Name`),
                      Signature = manypkgs::standardise_dates(`Date.of.Signature`),
                      Force = manypkgs::standardise_dates(`Date.of.Entry.into.Force`)) %>%
   dplyr::mutate(Beg = dplyr::coalesce(Signature, Force)) %>%
-  dplyr::select(GPTAD_ID, Country_ID, Country, Title, Beg, Signature, Force) %>%
+  dplyr::select(gptadID, CountryID, CountryName, Title, Beg, Signature, Force) %>%
   dplyr::arrange(Beg)
 
-#Add treaty_ID column
-GPTAD_MEM$treaty_ID <- manypkgs::code_agreements(GPTAD_MEM, GPTAD_MEM$Title, 
-                                                 GPTAD_MEM$Beg)
+#Add treatyID column
+GPTAD_MEM$treatyID <- manypkgs::code_agreements(GPTAD_MEM, GPTAD_MEM$Title, 
+                                                GPTAD_MEM$Beg)
 
-# Add many_ID column
-many_ID <- manypkgs::condense_agreements(manytrade::agreements,
-                                         var = c(DESTA$treaty_ID, GPTAD$treaty_ID,
-                                                 LABPTA$treaty_ID, TREND$treaty_ID))
-GPTAD_MEM <- dplyr::left_join(GPTAD_MEM, many_ID, by = "treaty_ID")
+# Add manyID column
+manyID <- manypkgs::condense_agreements(manytrade::memberships,
+                                        var = c(DESTA_MEM$treatyID, 
+                                                GPTAD_MEM$treatyID))
+GPTAD_MEM <- dplyr::left_join(GPTAD_MEM, manyID, by = "treatyID")
 
 # Re-order the columns
-GPTAD_MEM <- dplyr::relocate(GPTAD_MEM, many_ID)
+GPTAD_MEM <- dplyr::relocate(GPTAD_MEM, manyID, CountryID, Title, Beg, 
+                             Signature, Force, CountryName, gptadID)
+
+# Check for duplicates in manyID
+# duplicates <- GPTAD_MEM %>%
+#   dplyr::mutate(duplicates = duplicated(GPTAD_MEM[, c(1,2)])) %>%
+#   dplyr::relocate(manyID, CountryID,  duplicates)
+
+# delete rows that only have diff title but same Beg and other variables
+GPTAD_MEM <- subset(GPTAD_MEM, subset = !duplicated(GPTAD_MEM[, c(1,2,4,7,9)]))
 
 # manypkgs includes several functions that should help cleaning
 # and standardising your data.