+++ tests/testthat/test_GPTAD_MEM.R
@@ -1,4 +1,4 @@
-# Test if  meets the many universe requirements
+# Test if the dataset meets the many packages universe requirements
 
 # Report missing values
 test_that("missing observations are reported correctly", {
@@ -10,10 +10,9 @@ test_that("missing observations are reported correctly", {
   expect_false(any(grepl("n\\.a\\.$", memberships[["GPTAD_MEM"]])))
 })
 
-# Uniformity tests (agreements have a source ID, a string title, a signature and
-# entry into force date)
+# Uniformity tests (agreements have a countryID and Beg columns)
 test_that("datasets have the required variables", {
-  expect_col_exists(memberships[["GPTAD_MEM"]], vars(Country))
+  expect_col_exists(memberships[["GPTAD_MEM"]], vars(CountryID))
   expect_col_exists(memberships[["GPTAD_MEM"]], vars(Beg))
 })