+++ NEWS.md
@@ -1,3 +1,24 @@
+# manytrade 0.1.1
+
+2022-03-23
+
+## Package
+
+* Closed #19 by merging `manytrade::code_countryname()` functionality into `manystates::code_states()`.
+* Fixed #21 by redirecting reference to master branch to main branch in pushrelease workflow.
+* Closed #14 by adding codefactor and lintr to package workflows and as badges.
+
+## Data
+
+* Fixed #23 by amending regex list stored in package and modifying `manystates::code_states()` to return original names if unmatched in regex list.
+* Fixed #22 and #24 by renaming variables in databases.
+  * Renamed "treaty_ID" and "many_ID" across all databases to "treatyID" and "manyID" respectively.
+  * Renamed "L", "D", and "J" in the agreements database "DocType", "AgreementType", and "GeogArea" respectively.
+  * Reformatted "Beg", "Signature", "Force" variables from YYYY format to YYYY-MM-DD format in `DESTA`, `TREND`, `LABPTA` datasets to be consistent across the `agreements` database.
+  * Recoded `DocType` variable in `GPTAD` dataset to remove ambiguity in coding
+  * Added `GeogArea` variable in `GPTAD` dataset to remove
+  * Checked for duplicate entries in `agreements` and `memberships` datasets using `manyID`, `treatyID`, `CountryID`, `CountryName` and `Beg` variables and removed duplicate entries
+
 # manytrade 0.1.0
 
 2021-12-03